i2c: minimal driver, support irq interface
--------

+++ r4l/src/platform/platform_dev.rs
@@ -21,7 +21,7 @@ impl PlatformDevice {
 
 impl PlatformDevice {
     /// Returns irq of the platform device.
-    pub fn irq_resource(&self, index: u32) -> Result<i32> {
+    pub fn irq_resource(&self, index: usize) -> Result<u32> {
         self.device.irq_resource(index)
     }
 }
